UIKit extension of DimensionAttribute and Size
